// src/app/Model/pappoint.ts
export interface appointmentmodel {
  name?: string;
  age?: number;
  email?: string;
  phone?: string;
  address1?: string;
  address2?: string;
  city?: string;
  state?: string;
  country?: string;
  pinCode?: any;
  maleTrainer?: string;
  femaleTrainer?: string;
  noPreference?: string;
  yesP?: string;
  noP?: string;
}